backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.10.0
pydantic-settings==2.6.0

# Supabase
supabase==2.11.0

# Redis
redis[hiredis]==5.2.0

# LLM
langchain==0.3.0
langchain-community==0.3.0
langgraph==0.2.0
ollama==0.4.0

# HTTP
httpx==0.28.0
aiohttp==3.11.0

# Scheduling (Phase 2+)
apscheduler==3.10.4

# Image processing (Phase 3)
Pillow==11.0.0

# RSS parsing (Phase 2)
feedparser==6.0.11

# CORS
python-multipart==0.0.12
